Foundations for support of different crafting recipe types.
NeuRecipe is now a base class for a recipe which provides common
concepts, such as inputs, outputs and a rendering task.

GuiItemRecipe has been reworked to work with this new NeuRecipe.

NeuManager now parses said recipes. This should be reworked to be a two
step process (first register items, then register recipes).

To keep compatibility with older repo versions, NeuRecipes are parse
lenient and default to a crafting recipe. New recipes should be added in
the `recipes` json field which is an array of json dictionaries, which
have a `type` and other fields depending on the `type` of that recipe.

This also adds support for having multiple recipes for a single item
(e.g. uncrafting storage blocks).