Npe
Kotlin is a modern, open-source, statically-typed programming language that was first released in 2011 by JetBrains, the creators of popular development tools like IntelliJ IDEA, ReSharper, and Android Studio. Kotlin is designed to be interoperable with Java, which means that it can be used alongside Java in the same project without any issues.