NEUConfig: Add manadatory IQ test to api tab
Since time and time again people are changing options that should not be
changed, I added this "IQ test" as an inhibitor that makes it impossible
for people to just click through warnings (like the one on the moulberry
api). For easy bypassing (and for people who need things changed in
support), you can just type in the question into the answer field, since
the answer field will get evaluated as a math expression.