Add Tablist API
Todo: Add patterns for all the other widgets.

An Example implementation can be seen in MiningOverlay.java. If the requested Widget is not present, a TablistTutorial task will be added to the queue, which informs the user that they need to enable it.

Otherwise (hopefully) only the requested information will be returned, removing the need to constantly check which part of the tab list is currently being parsed.