Fix warning showing right after enabling widget & other todo items
Fix warning showing right after enabling the widget
This fix assumes the skyblock player is capable of not disabling the widget after enabling it.

Fix /neustats Has Advanced Tab
Instead added /neustats tablist, which copies the entire Tablist to the clipboard.
Deleted all SBInfo#hasNewTab usages in the process.

Also converted the MiningOverlay to show the warning inside the overlay, and only if the relevant lines are actually enabled.